feat: add ipython notebook for feature selection and outlier detection.
Preprocessed the data before pass to the machine learning
model.

feature_selection.ipynb is used to delete features
with a lot of missing values.
outlier_detection.ipynb is used to detect outliers
based on a simple prediction. Any values show a huge
difference between the simple prediction and their
actual values will be detected as outliers.